fixed dependency
Hi,

I'm trying to run your animation-timer code 'server' side in a nodejs module. All is fine, but there are two dependencies that don't match in two sub node_modules.

One in gm-parse-duration and the other one in gm-is. Both are very minor tweaks and when applied, the animation-timer runs like a charm.

thanks.